+++ orca_core/core.py
@@ -16,8 +16,7 @@ from threading import RLock
 import numpy as np
 from .hardware.dynamixel_client import DynamixelClient
 from .hardware.mock_dynamixel_client import MockDynamixelClient
-from .utils.yaml_utils import read_yaml, update_yaml # Changed from import *
-from .utils.load_utils import get_model_path
+from .utils.utils import *
 
 class OrcaHand:
     """OrcaHand class is used to abtract hardware control the hand of the robot with simple high level control methods in joint space."""
@@ -411,12 +410,14 @@ class OrcaHand:
         
         return overall_calibrated
 
-    def calibrate(self):
-        """Calibrate the hand by moving the joints to their limits and setting the ROMs. 
-        
-        The proecess is mostly hardware independent and is defined in the config.yaml file.
-        The motor position is increseed and decreased flexing and extending each joint while recording the reached limits.
-        """        
+    def calibrate(self, blocking: bool = True):
+        if blocking:
+            self._calibrate()
+        else:
+            self._start_task(self._calibrate)
+
+    def _calibrate(self):
+            
         # Store the min and max values for each motor
         motor_limits = self.motor_limits_dict.copy()
 
@@ -433,9 +434,15 @@ class OrcaHand:
         self.enable_torque()
         
         for step in self.calib_sequence:
+            if self._task_stop_event.is_set():
+                return
+
             desired_increment, motor_reached_limit, directions, position_buffers, motor_reached_limit, calibrated_joints, position_logs, current_log = {}, {}, {}, {}, {}, {}, {}, {}
 
             for joint, direction in step["joints"].items(): 
+                if self._task_stop_event.is_set():
+                    return
+
                 if joint == 'wrist':
                     self.set_max_current(self.wrist_calib_current)
                 else:
@@ -451,7 +458,7 @@ class OrcaHand:
                 current_log[motor_id] = []
                 motor_reached_limit[motor_id] = False
             
-            while(not all(motor_reached_limit.values())):                
+            while(not all(motor_reached_limit.values()) and not self._task_stop_event.is_set()):               
                 for motor_id, reached_limit in motor_reached_limit.items():
                     if not reached_limit:
                         desired_increment[motor_id] = directions[motor_id] * self.calib_step_size
@@ -742,10 +749,8 @@ class OrcaHand:
             
             min_pos, max_pos = self.joint_roms_dict[joint_name]
             
-            # Clip the position if outside ROM and notify
             if pos < min_pos or pos > max_pos:
                 clipped_pos = max(min_pos, min(max_pos, pos))
-                # print(f"Clipping {joint_name} from {pos} to {clipped_pos} (ROM: {min_pos} to {max_pos})")
                 pos = clipped_pos
 
             if self.joint_inversion_dict.get(joint_name, False):
@@ -799,8 +804,11 @@ class OrcaHand:
                 self.calibrated = False
                 update_yaml(self.calib_path, 'calibrated', False)
 
-    def tension(self, move_motors: bool = False):
-        self._start_task(self._tension, move_motors)
+    def tension(self, move_motors: bool = False, blocking: bool = True):
+        if blocking:
+            self._tension(move_motors)
+        else:
+            self._start_task(self._tension, move_motors)
 
     def _tension(self, move_motors: bool = False):
         """Freeze the motors, so that the hand can be manually tensioned.
