--- orca_core/utils/yaml_utils.py
@@ -1,43 +1,0 @@
-# ==============================================================================
-# Copyright (c) 2025 ORCA
-#
-# This file is part of ORCA and is licensed under the MIT License.
-# You may use, copy, modify, and distribute this file under the terms of the MIT License.
-# See the LICENSE file at the root of this repository for full license information.
-# ==============================================================================
-
-import yaml
-import numpy as np
-
-def update_yaml(file_path, key, value):
-    """Reads a YAML file, updates a specific key, and writes it back."""
-    if isinstance(value, np.ndarray):
-        value = value.tolist()
-    
-    if isinstance(value, dict):
-        value = {k: (v.tolist() if isinstance(v, np.ndarray) else v) for k, v in value.items()}
-
-    try:
-        with open(file_path, 'r+') as file:
-            data = yaml.safe_load(file) or {} 
-            
-            new_data = {key: value} # Make the latest chnage in the yaml file the first entry
-            for existing_key, existing_value in data.items():
-                if existing_key != key:
-                    new_data[existing_key] = existing_value
-            
-            file.seek(0) 
-            yaml.dump(new_data, file, default_flow_style=False, sort_keys=False) 
-            file.truncate() 
-    except FileNotFoundError:
-        with open(file_path, 'w') as file:
-            yaml.dump({key: value}, file, default_flow_style=False, sort_keys=False)
-
-
-def read_yaml(file_path):
-    """Reads a YAML file and returns its content."""
-    try:
-        with open(file_path, 'r') as file:
-            return yaml.safe_load(file) or None
-    except FileNotFoundError:
-        return {}
